#24: fixed trying to schedule event for null nextMeeting #?: replace reminder Join button with voice channel mention #?: move getManagedChannels() into channelEngine
